+++ frontend/src/models/User.ts
@@ -1,0 +1,6 @@
+export default interface User {
+    uid: string;
+    given_name: string;
+    mail: string;
+    is_admin: boolean;
+}
