+++ frontend/src/queries/User.ts
@@ -5,8 +5,8 @@ import {
     type UseQueryReturnType,
     type UseMutationReturnType,
 } from "@tanstack/vue-query";
-import { getUser, type User } from "@/helpers/datafetchers/User";
-import { authorized_fetch } from "@/helpers/datafetchers";
+import type User from "@/models/User";
+import { getUser, toggleAdmin } from "@/services/user";
 
 function USER_QUERY_KEY(uid?: string): string[] {
     return uid ? ["user", uid] : ["user"];
@@ -16,10 +16,10 @@ export function useUserQuery(uid?: string): UseQueryReturnType<User, Error> {
     return useQuery<User, Error>({ queryKey: USER_QUERY_KEY(uid), queryFn: () => getUser(uid) });
 }
 
-export function useToggleAdminMutation(): UseMutationReturnType<User, Error, User, void> {
+export function useToggleAdminMutation(): UseMutationReturnType<void, Error, User, void> {
     const queryClient = useQueryClient();
     return useMutation({
-        mutationFn: () => authorized_fetch("/api/users/me", { method: "POST" }),
+        mutationFn: toggleAdmin,
         onMutate: async (user: User) => {
             await queryClient.cancelQueries({ queryKey: USER_QUERY_KEY() });
             // const previousUser = queryClient.getQueryData<User>(USER_QUERY_KEY());
